updating and debugging yaml definitions and implementing controllers, need to be checked

+++ swagger/spring-server/src/main/java/ch/heig/amt/gamification/repositories/BadgeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Labinot Rashiti on 14/12/18.
  */
-public interface BadgeRepository extends CrudRepository<BadgeEntity, Long>{
+public interface BadgeRepository extends CrudRepository<BadgeEntity, Integer>{
 
 }
